Move if_else special form from compute core to compute lib
